data/product-engineering/pyproject.toml
[tool.poetry]
name = "product-engineering"
version = "0.1.0"
description = ""
authors = ["Daniel Bessonov <danielbess16@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.6"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
python-multipart = "^0.0.17"
openai = "^1.58.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"